type hints, rearranged data tables
