+++ my_proof/proof.py
@@ -1,15 +1,13 @@
-import json
+
 import logging
 import os
-from typing import Dict, Any
+from typing import Dict, Any, List, Callable
 
 import requests
 import gc
 import re
 import pandas as pd
 import hashlib
-import math
-import json
 
 from my_proof.models.proof_response import ProofResponse
 from .verify import DbSNPHandler
@@ -38,7 +36,8 @@ class TwentyThreeWeFileScorer:
     valid_genotypes = set("ATCG-ID")
     valid_chromosomes = set([str(i) for i in range(1, 23)] + ["X", "Y", "MT"])
 
-    def __init__(self, input_data, config, requester, hasher):
+    def __init__(self, input_data: List[str], config: Dict[str, Any],
+                 requester: requests.Session, hasher: Callable[[], hashlib.sha256]):
         self.input_data = input_data
         self.profile_id = self.get_profile_id(input_data)
         self.config = config
@@ -49,57 +48,52 @@ class TwentyThreeWeFileScorer:
         self.hasher = hasher
 
     @staticmethod
-    def get_profile_id(input_data):
+    def get_profile_id(input_data: List[str]) -> str | None:
         file_content = "\n".join([d for d in input_data[:50]])
 
-        # Define the URL pattern you're looking for
         url_prefix = 'https://you.23andme.com/p/'
         url_suffix = '/tools/data/download/'
 
-        # Find the starting position of the URL pattern
         start_index = file_content.find(url_prefix)
         if start_index == -1:
-            return None  # URL not found
+            return None
 
-        # Find the ending position of the profile ID (end of the URL)
         end_index = file_content.find(url_suffix, start_index)
         if end_index == -1:
-            return None  # URL suffix not found
+            return None
 
-        # Extract the profile ID between the URL prefix and suffix
         profile_id_start = start_index + len(url_prefix)
         profile_id = file_content[profile_id_start:end_index]
 
-        # Return the profile ID if found, otherwise None
         if profile_id:
             return profile_id
         else:
             return None
 
-    def read_header(self):
+    def read_header(self) -> str:
         header_lines = []
         for line in self.input_data:
-            if line.startswith("#") or line.startswith("rsid"):  # Capture header lines
-                # Ignore timestamp and profile-specific URL lines using regex
+            if line.startswith("#") or line.startswith("rsid"):
                 if re.match(r"# This data file generated by 23andMe at:", line):
-                    continue  # Skip unique timestamp line
+                    continue
                 if re.match(r"# https://you\.23andme\.com/p/", line):
-                    continue  # Skip unique download link line
+                    continue
                 header_lines.append(line.strip())
             else:
-                break  # Stop reading after headers
+                break
         header = "\n".join(header_lines[1:])
         return header
 
-    def check_header(self):
+    def check_header(self) -> bool:
         file_header = self.read_header()
 
-        clean_template = "\n".join([line.strip() for line in self.header_template.strip().split("\n") if """https://you.23andme.com""" not in line])
+        clean_template = "\n".join([line.strip() for line in self.header_template.strip().split("\n")
+                                    if """https://you.23andme.com""" not in line])
         clean_file_header = file_header.strip()
 
         return clean_template == clean_file_header
 
-    def check_rsid_lines(self):
+    def check_rsid_lines(self) -> bool:
         invalid_rows = []
 
         line_number = 1
@@ -119,17 +113,14 @@ class TwentyThreeWeFileScorer:
             rsid, chromosome, position, genotype = columns
             row = (rsid, chromosome, position, genotype)
 
-            # Check if rsid starts with 'rs' or 'i' and ends in digits
             if not re.match(r'^(rs|i)\d+$', rsid):
                 print(f"Line {line_number}: Invalid rsid format: {rsid}")
                 invalid_rows.append(row)
 
-            # Check if chromosome is one of 1-23, X, Y, or MT
             if chromosome not in self.valid_chromosomes:
                 print(f"Line {line_number}: Invalid chromosome value: {chromosome}")
                 invalid_rows.append(row)
 
-            # Check if the genotype contains only valid characters
             if any(char not in self.valid_genotypes for char in genotype):
                 print(f"Line {line_number}: Invalid genotype characters: {genotype}")
                 invalid_rows.append(row)
@@ -139,7 +130,7 @@ class TwentyThreeWeFileScorer:
             return False
         return True
 
-    def verify_profile(self):
+    def verify_profile(self) -> bool:
         """
         Sends the profile id for verification via a POST request.
         """
@@ -151,7 +142,7 @@ class TwentyThreeWeFileScorer:
 
         return profile_verified
 
-    def verify_hash(self, genome_hash):
+    def verify_hash(self, genome_hash: str) -> bool:
         """
         Sends the hashed genome data for verification via a POST request.
         """
@@ -162,26 +153,20 @@ class TwentyThreeWeFileScorer:
 
         return hash_unique
 
-    def hash_23andme_file(self, file_path):
-        # Read the 23andMe file into a DataFrame, skipping the comment lines
+    def hash_23andme_file(self, file_path: str) -> str:
         df = pd.read_csv(file_path, sep='\t', comment='#', names=['rsid', 'chromosome', 'position', 'genotype'])
 
-        # Remove rows where the rsid starts with 'i'
         df_filtered = df[~df['rsid'].str.startswith('i')]
 
-        # Sort the DataFrame by the rsid column
         df_sorted = df_filtered.sort_values(by='rsid')
 
-        # Concatenate the rows into one large symbol-separated string
         concatenated_string = '|'.join(
             df_sorted.apply(lambda row: f"{row['rsid']}:{row['chromosome']}:{row['position']}:{row['genotype']}",
                             axis=1))
 
-        # Delete the DataFrame to free up memory
         del df, df_filtered, df_sorted
-        gc.collect()  # Force garbage collection
+        gc.collect()
 
-        # Hash the concatenated string using SHA-256
         hash_object = self.hasher(concatenated_string.encode())
         hash_hex = hash_object.hexdigest()
         self.hash = hash_hex
@@ -189,7 +174,7 @@ class TwentyThreeWeFileScorer:
         return hash_hex
 
     @staticmethod
-    def invalid_genotypes_score(total: int, low: int = 1, high: int = 3):
+    def invalid_genotypes_score(total: int, low: int = 1, high: int = 3) -> float:
         if total <= low:
             return 1.0
         elif total >= high:
@@ -198,7 +183,7 @@ class TwentyThreeWeFileScorer:
             return 1.0 - (total - low) / (high - low)
 
     @staticmethod
-    def indel_score(total, low: int = 3, ultra_low: int = 1, high: int = 13, ultra_high: int = 22):
+    def indel_score(total: int, low: int = 3, ultra_low: int = 1, high: int = 13, ultra_high: int = 22) -> float:
         if total <= ultra_low:
             return 0.0
         elif ultra_low < total <= low:
@@ -211,7 +196,7 @@ class TwentyThreeWeFileScorer:
             return 0.0
 
     @staticmethod
-    def i_rsid_score(total: int, low: int = 5, high: int = 25):
+    def i_rsid_score(total: int, low: int = 5, high: int = 25) -> float:
         if total <= low:
             return 1.0
         elif total >= high:
@@ -221,7 +206,7 @@ class TwentyThreeWeFileScorer:
 
     @staticmethod
     def percent_verification_score(verified: int, all: int, low: float = 0.9, ultra_low: float = 0.85,
-                                   high: float = 0.96, ultra_high: float = 0.98):
+                                   high: float = 0.96, ultra_high: float = 0.98) -> float:
         verified_ratio = verified / all
 
         if low <= verified_ratio <= high:
@@ -235,7 +220,7 @@ class TwentyThreeWeFileScorer:
         else:
             return 0.0
 
-    def save_hash(self, proof_response):
+    def save_hash(self, proof_response: ProofResponse) -> bool:
 
         hash_data = self.generate_hash_save_data(proof_response)
         response = requests.post(url=self.config['key'], data=hash_data)
@@ -244,7 +229,7 @@ class TwentyThreeWeFileScorer:
 
         return success
 
-    def generate_hash_save_data(self, proof_response):
+    def generate_hash_save_data(self, proof_response: ProofResponse) -> Dict[str, Any]:
         hash_save_data = {
             'sender_address': self.sender_address,
             'attestor_address': '',
@@ -271,7 +256,7 @@ class TwentyThreeWeFileScorer:
         else:
             return 0
 
-    def proof_of_quality(self, filepath) -> float:
+    def proof_of_quality(self, filepath: str) -> float:
 
         dbsnp = DbSNPHandler(self.config)
         results = dbsnp.dbsnp_verify(filepath)
@@ -285,7 +270,7 @@ class TwentyThreeWeFileScorer:
 
         return quality_score
 
-    def proof_of_uniqueness(self, filepath) -> float:
+    def proof_of_uniqueness(self, filepath: str) -> float:
 
         hashed_dna = self.hash_23andme_file(filepath)
         unique = self.verify_hash(hashed_dna)
@@ -334,19 +319,16 @@ class Proof:
         self.proof_response.authenticity = scorer.proof_of_authenticity()
         self.proof_response.quality = scorer.proof_of_quality(filepath=twenty_three_file)
 
-        # Calculate overall score and validity
         total_score = (0.25 * self.proof_response.quality + 0.25 * self.proof_response.ownership +
                        0.25 * self.proof_response.authenticity + 0.25 * self.proof_response.uniqueness)
         self.proof_response.score = total_score
         self.proof_response.valid = total_score >= score_threshold
 
-        # Additional (public) properties to include in the proof about the data
         self.proof_response.attributes = {
             'total_score': total_score,
             'score_threshold': score_threshold,
         }
 
-        # Additional metadata about the proof, written onchain
         self.proof_response.metadata = {
             'dlp_id': self.config['dlp_id'],
         }
