+++ tests/test_proof.py
@@ -1,23 +1,40 @@
 
 import pytest
 from unittest import mock
-from unittest.mock import MagicMock, patch, Mock
+from unittest.mock import Mock, MagicMock
 from io import StringIO
 import pandas as pd
+from dataclasses import dataclass
+from typing import List, Dict, Optional
 
 from my_proof.proof import Proof, TwentyThreeWeFileScorer
 from my_proof.verify import DbSNPHandler
 
-from .conftest import mock_requester
 
-from .data_tables import (ProfileIdInputData, HeaderTestData, RSIDTestData, ProfileVerifyTestData, HashVerifyTestData,
-                          InvalidGenotypesScoreTestData, IndelScoreTestData, IRsidScoreTestData, ProofResponse,
-                          PercentVerifyScoreTestData, HashSaveDataTestData, ProofOfOwnershipTestData,
-                          ProofOfQualityTestData, ProofOfUniquenessTestData, ProofOfAuthenticityTestData,
-                          ReadHeaderTestData, Hash23AndMeFileTestData, SaveHashTestData, GenerateTestData)
+@pytest.fixture
+def mock_requester() -> mock.MagicMock:
+    """Fixture to mock requests."""
+    mock = MagicMock()
+    return mock
+
+
+@dataclass
+class ProofResponse:
+    authenticity: float
+    ownership: float
+    uniqueness: float
+    quality: float
+    attributes: Dict[str, float]
+    valid: bool
+    score: Optional[float]
+    dlp_id: Optional[int]
 
 
 class TestTwentyThreeWeFileScorer:
+    @dataclass
+    class ProfileIdInputData:
+        data: List[str]
+        expected_profile_id: Optional[str]
 
     @pytest.mark.parametrize("input_data", [
         ProfileIdInputData(data=[
@@ -33,9 +50,15 @@ class TestTwentyThreeWeFileScorer:
             "rsid\tchromosome\tposition\tgenotype"
         ], expected_profile_id="xyz456")
     ])
-    def test_get_profile_id(self, input_data):
+    def test_get_profile_id(self, input_data: ProfileIdInputData) -> None:
         scorer = TwentyThreeWeFileScorer(input_data.data, config={}, requester=None, hasher=None)
         assert scorer.get_profile_id(input_data.data) == input_data.expected_profile_id
+
+    @dataclass
+    class HeaderTestData:
+        data: List[str]
+        header_included: bool
+        expected_check_header: bool
 
     @pytest.mark.parametrize("input_data", [
         HeaderTestData(data=[
@@ -66,10 +89,15 @@ class TestTwentyThreeWeFileScorer:
             "rsid\tchromosome\tposition\tgenotype"
         ], header_included=False, expected_check_header=False)
     ])
-    def test_check_header(self, input_data):
+    def test_check_header(self, input_data: HeaderTestData) -> None:
         scorer = TwentyThreeWeFileScorer(input_data.data, config={}, requester=None, hasher=None)
         result = scorer.check_header()
         assert result == input_data.expected_check_header
+
+    @dataclass
+    class RSIDTestData:
+        data: List[str]
+        expected_check_rsid_lines: bool
 
     @pytest.mark.parametrize("input_data", [
         RSIDTestData(data=[
@@ -86,10 +114,16 @@ class TestTwentyThreeWeFileScorer:
             "rs8965392\tY\t54321\tINVALID_GENOTYPE"
         ], expected_check_rsid_lines=False)
     ])
-    def test_check_rsid_lines(self, input_data):
+    def test_check_rsid_lines(self, input_data: RSIDTestData) -> None:
         scorer = TwentyThreeWeFileScorer(input_data.data, config={}, requester=None, hasher=None)
         result = scorer.check_rsid_lines()
         assert result == input_data.expected_check_rsid_lines
+
+    @dataclass
+    class ProfileVerifyTestData:
+        profile_id: str
+        mock_response: Dict[str, bool]
+        expected_verification: bool
 
     @pytest.mark.parametrize("input_data", [
         ProfileVerifyTestData(profile_id="abc123", mock_response={'is_approved': True},
@@ -97,7 +131,7 @@ class TestTwentyThreeWeFileScorer:
         ProfileVerifyTestData(profile_id="xyz456", mock_response={'is_approved': False},
                               expected_verification=False)
     ])
-    def test_verify_profile(self, input_data, mock_requester):
+    def test_verify_profile(self, input_data: ProfileVerifyTestData, mock_requester: mock.MagicMock) -> None:
         config = {'verify': f"https://example.com/verify?address=abc"}
         input_data_lines = [f"# https://you.23andme.com/p/{input_data.profile_id}/tools/data/download/"]
         mock_requester.get.return_value.json.return_value = input_data.mock_response
@@ -105,113 +139,111 @@ class TestTwentyThreeWeFileScorer:
         scorer = TwentyThreeWeFileScorer(input_data_lines, config, requester=mock_requester, hasher=None)
         assert scorer.verify_profile() == input_data.expected_verification
 
+    @dataclass
+    class HashVerifyTestData:
+        genome_hash: str
+        mock_response: Dict[str, bool]
+        expected_hash_verification: bool
+
     @pytest.mark.parametrize("input_data", [
         HashVerifyTestData(genome_hash="abc_hash", mock_response={'is_unique': True},
-                                 expected_hash_verification=True),
+                           expected_hash_verification=True),
         HashVerifyTestData(genome_hash="xyz_hash", mock_response={'is_unique': False},
-                                 expected_hash_verification=False)
+                           expected_hash_verification=False)
     ])
-    def test_verify_hash(self, input_data, mock_requester):
+    def test_verify_hash(self, input_data: HashVerifyTestData, mock_requester: mock.MagicMock) -> None:
         config = {'key': "https://example.com/key"}
         mock_requester.get.return_value.json.return_value = input_data.mock_response
 
         scorer = TwentyThreeWeFileScorer(input_data=[], config=config, requester=mock_requester, hasher=None)
         assert scorer.verify_hash(input_data.genome_hash) == input_data.expected_hash_verification
 
-    # @pytest.mark.parametrize("input_data", [
-    #     Hash23AndMeFileTestData(
-    #         file_path="/mock/file/path",
-    #         mock_df_rows=[
-    #             {"rsid": "rs1", "chromosome": "chr1", "position": "pos1", "genotype": "AA"},
-    #             {"rsid": "rs2", "chromosome": "chr2", "position": "pos2", "genotype": "TT"}
-    #         ],
-    #         expected_concatenated_string="rs1:chr1:pos1:AA|rs2:chr2:pos2:TT",
-    #         expected_hash="hashed_value"
-    #     ),
-    #     Hash23AndMeFileTestData(
-    #         file_path="/another/mock/file",
-    #         mock_df_rows=[
-    #             {"rsid": "rs3", "chromosome": "chr3", "position": "pos3", "genotype": "GG"},
-    #             {"rsid": "rs4", "chromosome": "chr4", "position": "pos4", "genotype": "CC"}
-    #         ],
-    #         expected_concatenated_string="rs3:chr3:pos3:GG|rs4:chr4:pos4:CC",
-    #         expected_hash="another_hashed_value"
-    #     )
-    # ])
-    # @patch("pandas.read_csv")
-    # @patch("hashlib.sha256")
-    # def test_hash_23andme_file(self, mock_sha256, mock_read_csv, input_data):
-    #     """Test hash_23andme_file with a mocked pandas read_csv."""
-    #
-    #     mock_df = MagicMock()
-    #
-    #     mock_df.apply.return_value = [f"{row['rsid']}:{row['chromosome']}:{row['position']}:{row['genotype']}"
-    #                                   for row in input_data.mock_df_rows]
-    #     mock_read_csv.return_value = mock_df
-    #
-    #     mock_hash_object = MagicMock()
-    #     mock_hash_object.hexdigest.return_value = input_data.expected_hash
-    #     mock_sha256.return_value = mock_hash_object
-    #
-    #     scorer = TwentyThreeWeFileScorer(input_data=[], config={}, requester=None, hasher=mock_sha256)
-    #     result = scorer.hash_23andme_file(input_data.file_path)
-    #
-    #     mock_read_csv.assert_called_once_with(input_data.file_path, sep='\t', comment='#',
-    #                                           names=['rsid', 'chromosome', 'position', 'genotype'])
-    #
-    #     mock_sha256.assert_called_once_with(input_data.expected_concatenated_string.encode())
-    #     mock_hash_object.hexdigest.assert_called_once()
-    #
-    #     assert result == input_data.expected_hash
+    @dataclass
+    class InvalidGenotypesScoreTestData:
+        total: int
+        low: int
+        high: int
+        expected_score: float
 
     @pytest.mark.parametrize("input_data", [
         InvalidGenotypesScoreTestData(total=0, low=1, high=3, expected_score=1.0),
         InvalidGenotypesScoreTestData(total=2, low=1, high=3, expected_score=0.5),
         InvalidGenotypesScoreTestData(total=3, low=1, high=3, expected_score=0.0)
     ])
-    def test_invalid_genotypes_score(self, input_data):
+    def test_invalid_genotypes_score(self, input_data: InvalidGenotypesScoreTestData) -> None:
         result = TwentyThreeWeFileScorer.invalid_genotypes_score(input_data.total, input_data.low, input_data.high)
         assert result == input_data.expected_score
+
+    @dataclass
+    class IndelScoreTestData:
+        total: float
+        low: int
+        ultra_low: int
+        high: int
+        ultra_high: int
+        expected_score: float
 
     @pytest.mark.parametrize("input_data", [
         IndelScoreTestData(total=0, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=0.0),
         IndelScoreTestData(total=2, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=0.5),
         IndelScoreTestData(total=13, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=1.0),
-        IndelScoreTestData(total=18, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=4/9),
+        IndelScoreTestData(total=18, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=4 / 9),
         IndelScoreTestData(total=23, low=3, ultra_low=1, high=13, ultra_high=22, expected_score=0.0),
     ])
-    def test_indel_score(self, input_data):
+    def test_indel_score(self, input_data: IndelScoreTestData) -> None:
         result = TwentyThreeWeFileScorer.indel_score(input_data.total, input_data.low, input_data.ultra_low,
                                                      input_data.high, input_data.ultra_high)
         assert result == input_data.expected_score
+
+    @dataclass
+    class IRsidScoreTestData:
+        total: int
+        low: int
+        high: int
+        expected_score: float
 
     @pytest.mark.parametrize("input_data", [
         IRsidScoreTestData(total=0, low=5, high=25, expected_score=1.0),
         IRsidScoreTestData(total=15, low=5, high=25, expected_score=0.5),
         IRsidScoreTestData(total=30, low=5, high=25, expected_score=0.0)
     ])
-    def test_i_rsid_score(self, input_data):
+    def test_i_rsid_score(self, input_data: IRsidScoreTestData) -> None:
         result = TwentyThreeWeFileScorer.i_rsid_score(input_data.total, input_data.low, input_data.high)
         assert result == input_data.expected_score
 
+    @dataclass
+    class PercentVerifyScoreTestData:
+        verified: int
+        all: int
+        low: float
+        ultra_low: float
+        high: float
+        ultra_high: float
+        expected_score: float
+
     @pytest.mark.parametrize("input_data", [
         PercentVerifyScoreTestData(verified=90, all=100, low=0.9, ultra_low=0.85, high=0.96, ultra_high=0.98,
-                                         expected_score=1.0),
+                                   expected_score=1.0),
         PercentVerifyScoreTestData(verified=80, all=100, low=0.9, ultra_low=0.85, high=0.96, ultra_high=0.98,
-                                         expected_score=0.0),
+                                   expected_score=0.0),
         PercentVerifyScoreTestData(verified=85, all=100, low=0.9, ultra_low=0.85, high=0.96, ultra_high=0.98,
-                                         expected_score=0.0),
+                                   expected_score=0.0),
         PercentVerifyScoreTestData(verified=97, all=100, low=0.9, ultra_low=0.85, high=0.96, ultra_high=0.98,
-                                         expected_score=0.5),
+                                   expected_score=0.5),
         PercentVerifyScoreTestData(verified=88, all=100, low=0.9, ultra_low=0.85, high=0.96, ultra_high=0.98,
-                                         expected_score=0.6),
+                                   expected_score=0.6),
     ])
-    def test_percent_verification_score(self, input_data):
+    def test_percent_verification_score(self, input_data: PercentVerifyScoreTestData) -> None:
         result = TwentyThreeWeFileScorer.percent_verification_score(
             input_data.verified, input_data.all, input_data.low,
             input_data.ultra_low, input_data.high, input_data.ultra_high
         )
         assert result == input_data.expected_score
+
+    @dataclass
+    class HashSaveDataTestData:
+        proof_response: Dict[str, any]
+        expected_hash_save_data: Dict[str, any]
 
     @pytest.mark.parametrize("input_data", [
         HashSaveDataTestData(
@@ -271,7 +303,7 @@ class TestTwentyThreeWeFileScorer:
             }
         )
     ])
-    def test_generate_hash_save_data(self, input_data):
+    def test_generate_hash_save_data(self, input_data: HashSaveDataTestData) -> None:
         scorer = TwentyThreeWeFileScorer(
             input_data=[],
             config={},
@@ -286,6 +318,11 @@ class TestTwentyThreeWeFileScorer:
 
         assert result == input_data.expected_hash_save_data
 
+    @dataclass
+    class ProofOfOwnershipTestData:
+        mock_verify_profile_response: bool
+        expected_ownership_score: float
+
     @pytest.mark.parametrize("input_data", [
         ProofOfOwnershipTestData(
             mock_verify_profile_response=True,
@@ -296,8 +333,8 @@ class TestTwentyThreeWeFileScorer:
             expected_ownership_score=0.0
         )
     ])
-    def test_proof_of_ownership(self, input_data, monkeypatch):
-        def mock_verify_profile(self):
+    def test_proof_of_ownership(self, input_data: ProofOfOwnershipTestData, monkeypatch: pytest.MonkeyPatch) -> None:
+        def mock_verify_profile(self) -> bool:
             return input_data.mock_verify_profile_response
 
         monkeypatch.setattr(TwentyThreeWeFileScorer, 'verify_profile', mock_verify_profile)
@@ -307,6 +344,11 @@ class TestTwentyThreeWeFileScorer:
         result = scorer.proof_of_ownership()
 
         assert result == input_data.expected_ownership_score
+
+    @dataclass
+    class ProofOfQualityTestData:
+        dbsnp_verify_result: Dict[str, int]
+        expected_quality_score: float
 
     @pytest.mark.parametrize("input_data", [
         ProofOfQualityTestData(
@@ -330,7 +372,7 @@ class TestTwentyThreeWeFileScorer:
             expected_quality_score=0.25
         )
     ])
-    def test_proof_of_quality(self, input_data, mocker):
+    def test_proof_of_quality(self, input_data: ProofOfQualityTestData, mocker: mock.MagicMock) -> None:
         sample_data = """rsid	chromosome	position	genotype
         rs137900170	1	1392325	GG
         rs138988486	1	1394069	CC
@@ -360,10 +402,17 @@ class TestTwentyThreeWeFileScorer:
 
         mocker.patch("my_proof.verify.DbSNPHandler", return_value=mock_dbsnp_handler)
 
-        scorer = TwentyThreeWeFileScorer(input_data=[], config={'token': 'test-token', 'endpoint': 'test-endpoint'}, requester=None, hasher=None)
+        scorer = TwentyThreeWeFileScorer(input_data=[], config={'token': 'test-token', 'endpoint': 'test-endpoint'},
+                                         requester=None, hasher=None)
         result = scorer.proof_of_quality(filepath="dummy_filepath")
 
         assert result == pytest.approx(input_data.expected_quality_score, rel=1e-2)
+
+    @dataclass
+    class ProofOfUniquenessTestData:
+        mock_hash_23andme_file_response: str
+        mock_verify_hash_response: bool
+        expected_uniqueness_score: float
 
     @pytest.mark.parametrize("input_data", [
         ProofOfUniquenessTestData(
@@ -377,13 +426,19 @@ class TestTwentyThreeWeFileScorer:
             expected_uniqueness_score=0.0
         )
     ])
-    def test_proof_of_uniqueness(self, input_data, mocker):
+    def test_proof_of_uniqueness(self, input_data: ProofOfUniquenessTestData, mocker: mock.MagicMock) -> None:
         scorer = TwentyThreeWeFileScorer(input_data=[], config={}, requester=None, hasher=None)
         mocker.patch.object(scorer, 'hash_23andme_file', return_value=input_data.mock_hash_23andme_file_response)
         mocker.patch.object(scorer, 'verify_hash', return_value=input_data.mock_verify_hash_response)
         result = scorer.proof_of_uniqueness(filepath="dummy_path")
 
         assert result == input_data.expected_uniqueness_score
+
+    @dataclass
+    class ProofOfAuthenticityTestData:
+        mock_check_header_response: bool
+        mock_check_rsid_lines_response: bool
+        expected_authenticity_score: float
 
     @pytest.mark.parametrize("input_data", [
         ProofOfAuthenticityTestData(
@@ -407,13 +462,18 @@ class TestTwentyThreeWeFileScorer:
             expected_authenticity_score=0.0
         )
     ])
-    def test_proof_of_authenticity(self, input_data, mocker):
+    def test_proof_of_authenticity(self, input_data: ProofOfAuthenticityTestData, mocker: mock.MagicMock) -> None:
         scorer = TwentyThreeWeFileScorer(input_data=[], config={}, requester=None, hasher=None)
         mocker.patch.object(scorer, 'check_header', return_value=input_data.mock_check_header_response)
         mocker.patch.object(scorer, 'check_rsid_lines', return_value=input_data.mock_check_rsid_lines_response)
         result = scorer.proof_of_authenticity()
 
         assert result == input_data.expected_authenticity_score
+
+    @dataclass
+    class ReadHeaderTestData:
+        input_data: list[str]
+        expected_header: str
 
     @pytest.mark.parametrize("input_data", [
         ReadHeaderTestData(
@@ -473,12 +533,18 @@ class TestTwentyThreeWeFileScorer:
 # rsid\tchromosome\tposition\tgenotype"""
         ),
     ])
-    def test_read_header(self, input_data):
+    def test_read_header(self, input_data: ReadHeaderTestData) -> None:
         scorer = TwentyThreeWeFileScorer(input_data=input_data.input_data, config={}, requester=None, hasher=None)
 
         result = scorer.read_header()
 
         assert result == input_data.expected_header
+
+    @dataclass
+    class Hash23AndMeFileTestData:
+        input_data: pd.DataFrame
+        expected_concatenated_string: str
+        expected_hash: str
 
     @pytest.mark.parametrize("input_data", [
         Hash23AndMeFileTestData(
@@ -502,7 +568,7 @@ class TestTwentyThreeWeFileScorer:
             expected_hash="dummy_hash_2"
         )
     ])
-    def test_hash_23andme_file(self, input_data, mocker):
+    def test_hash_23andme_file(self, input_data: Hash23AndMeFileTestData, mocker: mock.MagicMock) -> None:
         mocker.patch('pandas.read_csv', return_value=input_data.input_data)
 
         mock_hasher = Mock()
@@ -520,6 +586,13 @@ class TestTwentyThreeWeFileScorer:
         )
 
         mock_hasher.assert_called_with(concatenated_string.encode())
+
+    @dataclass
+    class SaveHashTestData:
+        proof_response: dict
+        expected_hash_data: dict
+        mock_response: dict
+        expected_success: bool
 
     @pytest.mark.parametrize("input_data", [
         SaveHashTestData(
@@ -579,49 +652,51 @@ class TestTwentyThreeWeFileScorer:
             expected_success=False
         )
     ])
-    def test_save_hash(self, input_data, mocker):
-        # Mock generate_hash_save_data to return the expected hash data
+    def test_save_hash(self, input_data: SaveHashTestData, mocker: mock.MagicMock) -> None:
         scorer = TwentyThreeWeFileScorer(input_data=[], config={'key': 'https://example.com/save'}, requester=None,
                                          hasher=None)
         mocker.patch.object(scorer, 'generate_hash_save_data', return_value=input_data.expected_hash_data)
 
-        # Mock the requests.post to simulate the API call
         mock_post = mocker.patch('requests.post')
         mock_post.return_value.json.return_value = input_data.mock_response
 
-        # Mock other attributes
         scorer.sender_address = 'test_sender'
         scorer.profile_id = 'test_profile'
         scorer.hash = 'test_hash'
 
-        # Call the method
         result = scorer.save_hash(input_data.proof_response)
 
-        # Assert the request was made with the correct data
         mock_post.assert_called_with(url='https://example.com/save', data=input_data.expected_hash_data)
 
-        # Assert the result matches the expected success flag
         assert result == input_data.expected_success
 
 
 class TestProof:
 
+    @dataclass
+    class ProofGenerateTestData:
+        file_list: list
+        expected_uniqueness: float
+        expected_ownership: float
+        expected_authenticity: float
+        expected_quality: float
+        expected_valid: bool
+
     @pytest.mark.parametrize("test_data", [
-        GenerateTestData(file_list=['file1.txt'], expected_uniqueness=1.0, expected_ownership=1.0,
-                         expected_authenticity=1.0, expected_quality=1.0, expected_valid=True),
-        GenerateTestData(file_list=['file2.txt'], expected_uniqueness=0.9, expected_ownership=0.9,
-                         expected_authenticity=0.9, expected_quality=0.9, expected_valid=True),
-        GenerateTestData(file_list=['file3.txt'], expected_uniqueness=0.5, expected_ownership=0.5,
-                         expected_authenticity=0.5, expected_quality=0.5, expected_valid=False),
+        ProofGenerateTestData(file_list=['file1.txt'], expected_uniqueness=1.0, expected_ownership=1.0,
+                              expected_authenticity=1.0, expected_quality=1.0, expected_valid=True),
+        ProofGenerateTestData(file_list=['file2.txt'], expected_uniqueness=0.9, expected_ownership=0.9,
+                              expected_authenticity=0.9, expected_quality=0.9, expected_valid=True),
+        ProofGenerateTestData(file_list=['file3.txt'], expected_uniqueness=0.5, expected_ownership=0.5,
+                              expected_authenticity=0.5, expected_quality=0.5, expected_valid=False),
     ])
     @mock.patch('os.listdir')
     @mock.patch('builtins.open', new_callable=mock.mock_open, read_data="test data")
     @mock.patch('my_proof.proof.TwentyThreeWeFileScorer')
-    def test_generate(self, mock_scorer_cls, mock_open, mock_listdir, test_data: GenerateTestData):
-        # Mocking the list of input files
+    def test_generate(self, mock_scorer_cls: mock.MagicMock, mock_open: mock.MagicMock,
+                      mock_listdir: mock.MagicMock, test_data: ProofGenerateTestData) -> None:
         mock_listdir.return_value = test_data.file_list
 
-        # Mock the scorer instance
         mock_scorer = mock.Mock()
         mock_scorer.proof_of_uniqueness.return_value = test_data.expected_uniqueness
         mock_scorer.proof_of_ownership.return_value = test_data.expected_ownership
@@ -629,25 +704,19 @@ class TestProof:
         mock_scorer.proof_of_quality.return_value = test_data.expected_quality
         mock_scorer.save_hash.return_value = True
 
-        # Assign the mocked scorer to be returned when instantiating TwentyThreeWeFileScorer
         mock_scorer_cls.return_value = mock_scorer
 
-        # Create a mock ProofResponse object without dlp_id
         mock_proof_response = ProofResponse(score=0, authenticity=0, ownership=0, uniqueness=0, quality=0, valid=False,
                                             attributes={'score': 0}, dlp_id=1234)
 
-        # Set up the config and proof instance
         config = {'input_dir': '/input', 'dlp_id': 1234, 'key': 'some_key', 'verify': 'some_verify'}
         proof = Proof(config)
         proof.proof_response = mock_proof_response
 
-        # Call the method to test
         proof_response = proof.generate()
 
-        # Manually set the dlp_id in the proof_response (if needed)
         proof_response.metadata['dlp_id'] = config['dlp_id']
 
-        # Assert that the scorer methods were called
         mock_scorer.proof_of_uniqueness.assert_called_once_with(filepath=f'/input/{test_data.file_list[0]}')
         mock_scorer.proof_of_ownership.assert_called_once()
         mock_scorer.proof_of_authenticity.assert_called_once()
@@ -660,10 +729,6 @@ class TestProof:
         assert proof_response.attributes['total_score'] == total_score
         assert proof_response.attributes['score_threshold'] == 0.9
 
-        # Assert that save_hash was called
         mock_scorer.save_hash.assert_called_once_with(proof_response)
 
-        # Ensure the metadata is correct
         assert proof_response.metadata['dlp_id'] == 1234
-
-
